omapdss: dss_dpll: Don't use DSS_PLL_CONTROL to check if DSS PLLs are in use
We currently check if a DSS VIDEO/HDMI PLL is in use by reading the
PLL_x_DSS_CONTROL_DISABLE fields in DSS_PLL_COTNROL.

This register doesn't change it's state after a soft reboot, and hence may
not reflect correctly to the driver whether a PLL is in use or not. Use
software variables to keep a track of the PLL usage.

This prevents the situation where an interface finds all the PLLs in use and
uses DSS_FCLK as it's pixel clock source. Lowering DSS_FCLK might cause another
interface to fail to display correctly as DISPC may not be able to process
pixels for it with the lower functional clock. An example of this is on the
DRA7-EVM, where LCD reduces DSS_FCLK, causing HDMI to not function correctly
anymore.

Change-Id: I99ed7757eac79f7c568faeb98e281bdbd1ee4a33
Signed-off-by: Archit Taneja <archit@ti.com>