drm: warn when config FB is non-NULL
a check is put into set_config_internal which ensures
the fb sent as argument to set_config is the same as
crtc->fb after the function returns.

But this check is supposed to fail when the connector
is unplugged as the new crtc->fb is going to be NULL.

The new check also evaluates this condition.

Change-Id: I0069eada22e3f0a03342be2c65a3754fed926c3d
Signed-off-by: Subhajit Paul <a0132170@ti.com>