OMAPDSS: Displays: Add tlc59108 based DPI panel support
Add a panel driver for the DPI panels on DRA7 and DRA72 EVMS.

The DPI panels themselves could have been implemented by a generic DPI panel
driver. However, the EVMs have an intermediate chip "TLC59108" which controls
the backlight and power GPIOs for the panels.

This driver combines both the panel and TLC chip i2c drivers into one single
driver, and therefore isn't the correct implementation. The TLC chip should
ideally be a mfd driver functioning as a gpio expander and a pwm chip.

We add a mechanism of selecting a different set of video timings by matching it
with the compatible DT string. The timings are for the 800x480 panel on
DRA7-EVM, and the 1280x800 panel on DRA72x-EVM.

Signed-off-by: Archit Taneja <archit@ti.com>