OMAPDSS: Add a PLL library
The DSI, HDMI and the VIDEO PLLs(on DRA7x) are very similar in terms of register
programming model and PLL parameter calculations.

An overlay manager's pixel clock can be driven by DSI PLL on OMAP3, DSI and HDMI
PLLs on OMAP4/5, and VIDEO and HDMI PLL on DRA7x.

This can make the DPI driver quite messy if we want to use different PLL API for
differnt SoCs.

We create a pll library, which provides a way to register to a pll instance,
which an output driver(like DSI, DPI, HDMI) can use.

Currently done only for DSI/HDMI type PLLs.

Signed-off-by: Archit Taneja <archit@ti.com>