OMAPDSS: Add a custom TPD level shifter encoder driver for DRA7 EVM
DRA7 EVM has quite a complicated path between the HDMI DDC, HPD lines and the
TPD level shifter. This includes a demuxer which choses between I2C2 bus and
HDMI DDC bus. The I2C2 bus further drives GPIO expanders which are required for
configuring the level shifter input pins.

For correct operation of the HDMI level shifter and HDMI DDC, we need to enable
a switching logic, which enables HDMI DDC when edid needs to be read, and I2C2
when we need to read or configure level shifter pins.

The current display model cannot be used to implement such path. Implement a
custom encoder driver which drivers the GPIO to configure the demux, and
performs demux switching wherever needed.

Add hot plug detect support in the tpd encoder driver used by DRA7 EVM. Hot plug
detect on the dra7 evm became functional only after rev E. This revision has a
direct connection of the HPD_A signal from the tpd level shifter to a GPIO.

The previous EVMs had the signal going via a gpio expander, which forced us to
keep the I2C2/HDMI demux in HDMI mode whenever possible to detect a hot plug
event. We now remove the demuxing, and limit it only to reading EDID.

Add a hpd handler which configures the LS_OE line like done in the original
tpd encoder driver.

The HPD_B signal on the level shifter shows a glitch on cable connect. The TPD
level shifter hardware is supposed to prevent these glitches using it's own
debouncing hardware. But the glitch is seen on HPD_A too, and requires us to
set debounce for the DRA7x gpio.
The nature of the glitch is an instant rise to around 3V, followed by a lot of
noise, and ending with a slow taper down of the voltage. The actual connect
event comes around 80 ms later. Keep the omap-gpio debounce time to
1 millisecond to ensure that the complete glitch duration is covered. The
reason why the glitch isn't filtered by the level shifter needs to be
investigated.

Create a DT property called "disable-hpd" to disable hot plug detect. This
should be added in the encoder's DT node for pre-revE+ boards. Having this
property prevents the driver from registering an irq-handler, and performs a
manual enable of LS_OE pin on the TPD chip. pre-revE+ boards also require a
demux as the hpd signal reaches the gpio only when we switch to the HDMI demux
mode, preform that too when the property exists.

Signed-off-by: Archit Taneja <archit@ti.com>